# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A wrapper around ffmpeg",
        "title": "ffmate API",
        "contact": {
            "name": "We love media",
            "email": "sev@welovemedia.io"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://opensource.org/license/agpl-v3"
        }
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/watchfolder": {
            "put": {
                "description": "Update a watchfolder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchfolders"
                ],
                "summary": "Update a watchfolder",
                "parameters": [
                    {
                        "description": "new watchfolder",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewWatchfolder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Watchfolder"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new watchfolder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchfolders"
                ],
                "summary": "Add a new watchfolder",
                "parameters": [
                    {
                        "description": "new watchfolder",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewWatchfolder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Watchfolder"
                        }
                    }
                }
            }
        },
        "/watchfolder/{uuid}": {
            "get": {
                "description": "Get a single watchfolder by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchfolders"
                ],
                "summary": "Get single watchfolder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the watchfolders uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Watchfolder"
                        }
                    }
                }
            }
        },
...
        "/watchfolders/{uuid}": {
            "delete": {
                "description": "Delete a watchfolder by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchfolders"
                ],
                "summary": "Delete a watchfolder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the watchfolders uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/webhooks": {
            "get": {
                "description": "List all existing webhooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "List all webhooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Webhook"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new webhook for an event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Add a new webhook",
                "parameters": [
                    {
                        "description": "new webhook",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewWebhook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Webhook"
                        }
                    }
                }
            }
        },
        "/webhooks/{uuid}": {
            "get": {
                "description": "Get a single webhook by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Get single webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the webhooks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Webhook"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a webhook for an event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Update a webhook",
                "parameters": [
                    {
                        "description": "updated webhook",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewWebhook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Webhook"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a webhook by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the webhooks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.NewWatchfolder": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "filter": {
                    "$ref": "#/definitions/dto.WatchfolderFilter"
                },
                "growthChecks": {
                    "type": "integer"
                },
                "interval": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "preset": {
                    "type": "string"
                },
                "suspended": {
                    "type": "boolean"
                }
            }
        },
        "dto.NewWebhook": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/dto.WebhookEvent"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dto.Watchfolder": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "filter": {
                    "$ref": "#/definitions/dto.WatchfolderFilter"
                },
                "growthChecks": {
                    "type": "integer"
                },
                "interval": {
                    "type": "integer"
                },
                "lastCheck": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "preset": {
                    "type": "string"
                },
                "suspended": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.WatchfolderFilter": {
            "type": "object",
            "properties": {
                "extensions": {
                    "$ref": "#/definitions/dto.WatchfolderFilterExtensions"
                }
            }
        },
        "dto.WatchfolderFilterExtensions": {
            "type": "object",
            "properties": {
                "exclude": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.Webhook": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/dto.WebhookEvent"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.WebhookEvent": {
            "type": "string",
            "enum": [
                "batch.created",
                "batch.finished",
                "task.created",
                "task.updated",
                "task.deleted",
                "preset.created",
                "preset.updated",
                "preset.deleted",
                "webhook.created",
                "webhook.updated",
                "webhook.deleted",
                "watchfolder.created",
                "watchfolder.updated",
                "watchfolder.deleted"
            ],
            "x-enum-varnames": [
                "BATCH_CREATED",
                "BATCH_FINISHED",
                "TASK_CREATED",
                "TASK_UPDATED",
                "TASK_DELETED",
                "PRESET_CREATED",
                "PRESET_UPDATED",
                "PRESET_DELETED",
                "WEBHOOK_CREATED",
                "WEBHOOK_UPDATED",
                "WEBHOOK_DELETED",
                "WATCHFOLDER_CREATED",
                "WATCHFOLDER_UPDATED",
                "WATCHFOLDER_DELETED"
            ]
        }
    }
}